+++ cras/src/server/linear_resampler.h
@@ -17,9 +17,9 @@ struct linear_resampler;
  *    dst_rate - The destination rate to resample to.
  */
 struct linear_resampler *linear_resampler_create(unsigned int num_channels,
-					     unsigned int format_bytes,
-					     unsigned int src_rate,
-					     unsigned int dst_rate);
+						 unsigned int format_bytes,
+						 float src_rate,
+						 float dst_rate);
 
 /* Sets the rates for the linear resampler.
  * Args:
@@ -27,8 +27,8 @@ struct linear_resampler *linear_resampler_create(unsigned int num_channels,
  *    to - The rate to resample to.
  */
 void linear_resampler_set_rates(struct linear_resampler *lr,
-			      unsigned int from,
-			      unsigned int to);
+				float from,
+				float to);
 
 /* Converts the frames count from output rate to input rate. */
 unsigned int linear_resampler_out_frames_to_in(struct linear_resampler *lr,
